test: update recovery_key_service_test
