+++ apps/firmware_updater/test/recovery_key_service_test.dart
@@ -1,6 +1,9 @@
+import 'dart:io';
+
 import 'package:firmware_updater/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/annotations.dart';
+import 'package:mockito/mockito.dart';
 import 'package:snapd/snapd.dart';
 
 import 'test_utils.dart';
@@ -12,6 +15,7 @@ void main() {
     final service = RecoveryKeySnapdService(
       snapdClient: snapdClient,
       udisksClient: mockUDisksClient(),
+      runProcess: mockProcess().run,
     );
     expect(await service.checkRecoveryKey('foo'), isTrue);
   });
@@ -21,6 +25,7 @@ void main() {
     final service = RecoveryKeySnapdService(
       snapdClient: snapdClient,
       udisksClient: mockUDisksClient(),
+      runProcess: mockProcess().run,
     );
     expect(await service.checkRecoveryKey('foo'), isFalse);
   });
@@ -30,6 +35,7 @@ void main() {
     final service = RecoveryKeySnapdService(
       snapdClient: mockSnapdClient(),
       udisksClient: udisks,
+      runProcess: mockProcess().run,
     );
     expect(service.hasBitlocker, isTrue);
   });
@@ -38,43 +44,38 @@ void main() {
     final service = RecoveryKeySnapdService(
       snapdClient: mockSnapdClient(),
       udisksClient: udisks,
+      runProcess: mockProcess().run,
     );
     expect(service.hasBitlocker, isFalse);
   });
 
-  test('snapd fde active', () async {
-    final snapdClient = mockSnapdClient(
-      storageEncryptionStatus: SnapdStorageEncryptionStatus.active,
-    );
+  test('snapd fde managed', () async {
+    final snapdClient = mockSnapdClient();
     final service = RecoveryKeySnapdService(
       snapdClient: snapdClient,
       udisksClient: mockUDisksClient(),
+      runProcess: mockProcess(stdout: 'storage-encrypted: managed').run,
     );
     await service.init();
     expect(service.hasUbuntuFde, isTrue);
   });
 
-  test('snapd fde inactive', () async {
-    final snapdClient = mockSnapdClient(
-      storageEncryptionStatus: SnapdStorageEncryptionStatus.inactive,
-    );
+  test('snapd fde not detected', () async {
+    final snapdClient = mockSnapdClient();
     final service = RecoveryKeySnapdService(
       snapdClient: snapdClient,
       udisksClient: mockUDisksClient(),
+      runProcess: mockProcess().run,
     );
     await service.init();
     expect(service.hasUbuntuFde, isFalse);
   });
+}
 
-  test('snapd fde degraded', () async {
-    final snapdClient = mockSnapdClient(
-      storageEncryptionStatus: SnapdStorageEncryptionStatus.degraded,
-    );
-    final service = RecoveryKeySnapdService(
-      snapdClient: snapdClient,
-      udisksClient: mockUDisksClient(),
-    );
-    await service.init();
-    expect(service.hasUbuntuFde, isTrue);
-  });
+MockProcess mockProcess({String? stdout}) {
+  final mockProcess = MockProcess();
+  when(mockProcess.run('snapctl', ['system-mode'])).thenAnswer(
+    (_) async => ProcessResult(0, 0, stdout ?? '', ''),
+  );
+  return mockProcess;
 }
